Rename hook folder to create-use-styles and hoc to with-styles
