+++ packages/react-jss/src/with-styles.js
@@ -4,17 +4,17 @@ import hoistNonReactStatics from 'hoist-non-react-statics'
 import {type StyleSheet} from 'jss'
 import {ThemeContext} from 'theming'
 
-import type {HOCProps, HOCOptions, Styles, InnerProps, DynamicRules} from '../types'
+import type {HOCProps, HOCOptions, Styles, InnerProps, DynamicRules} from './types'
 import getDisplayName from './getDisplayName'
-import memoize from '../utils/memoize-one'
-import mergeClasses from '../utils/merge-classes'
-import {JssContext} from '../JssContext'
-import {getIndex} from '../utils/index-counter'
-import {createSheet} from '../utils/create-sheet'
-import {addDynamicRules, removeDynamicRules, updateDynamicRules} from '../utils/dynamic-rules'
-import {unmanageSheet} from '../utils/unmanage-sheet'
-import {manageSheet} from '../utils/manage-sheet'
-import {getSheetClasses} from '../utils/get-sheet-classes'
+import memoize from './utils/memoize-one'
+import mergeClasses from './utils/merge-classes'
+import {JssContext} from './JssContext'
+import {getIndex} from './utils/index-counter'
+import {createSheet} from './utils/create-sheet'
+import {addDynamicRules, removeDynamicRules, updateDynamicRules} from './utils/dynamic-rules'
+import {unmanageSheet} from './utils/unmanage-sheet'
+import {manageSheet} from './utils/manage-sheet'
+import {getSheetClasses} from './utils/get-sheet-classes'
 
 interface State {
   dynamicRules: ?DynamicRules;
